+++ src/level/chunk/mod.rs
@@ -1,10 +1,17 @@
 // Copyright 2025 Gabriel Bjørnager Jensen.
 
-use crate::level::Material;
+use crate::level::Layer;
 
+use std::borrow::Cow;
+
+/// A level chunk.
+///
+/// Levels are made up of a non-zero amount of segments that are called *chunks*.
+/// Each chunk defines additional *layers* itself (see [`Layer`]).
 #[derive(Clone, Debug)]
 pub struct Chunk {
-	pub terrain_height: f64,
+	/// The width of the chunk, relative to the map width.
+	pub width: f64,
 
-	pub ground: Material,
+	pub layers: Cow<'static, [Layer]>,
 }