+++ src/level/mod.rs
@@ -3,6 +3,7 @@
 mod block;
 mod block_tags;
 mod chunk;
+mod layer;
 mod level;
 mod material;
 mod material_from_str_error;
@@ -10,6 +11,7 @@ mod material_from_str_error;
 pub use block::Block;
 pub use block_tags::BlockTags;
 pub use chunk::Chunk;
+pub use layer::Layer;
 pub use level::Level;
 pub use material::Material;
 pub use material_from_str_error::MaterialFromStrError;